Revert "1 slight refactor"
