--- .gitignore
@@ -1,5 +1,0 @@
-*.pyc
-__pycache__/
-.venv/
-*.zip
-.coverage
